rutabaga_gfx: move RutabagaHandle to it's own file
- Add as_mesa_handle() function which cleans up handle access
- Move handle logic to it's own file instead of rutabaga_core
- Other minor improvements
